This commit makes an initial set of utilities for StreamCorpus.
As noted in the comments of the Makefile, this pushes the results of
thrift code generation into our python package.

	## The 'right' thing to do here is apparently to use the
	## 'namespace' functionality in python setup tools.  However,
	## this is not yet adopted into core python and has some
	## kludgy sounding stuff about using __init__.py for the sole
	## purpose of causing (tricking?) python to looking further
	## through sys.path if it fails to find a component in the
	## first module with the requested name.  Instead of taking a
	## depency on that, we maintain control of the kludge here in
	## this Makefile by copying the code files generated by
	## thrift into our working tree and not the empty __init__.py
	## files from thrift.

The setup.py successfully builds and installs the StreamCorpus
package, however it does not yet create RPMs.  Looks like need to add
this to Makefile to make cleanup repeatable.

The main content in this commit is in src/StreamCorpus/__init__.py and
src/StreamCorpus.thrift which has significant changes compared with
the original kba.thrift design.  Moved the non-body ContentItems into
a dict, and similarly moved SourceMetadata into a dict.

The __init__.py file in StreamCorpus introduces the notion of a Chunk
and offers convenience methods on chunks.  This is not yet totally
pythonic but getting closer.

This does not yet exercise Labels and Offsets.  That's part of what is
next.