Fix build from sdist
We currently assume that the thrift generated python files are included in our source
generation.  If we actually want other users to be able to generate the thrift files (
which is very reasonable) we need to include the interface files in the
source distribution.  Comment out thrift generation as part of 'python
setup.py build' until them.