Exit with a non-zero value if any of the tests fail
Buildbot was passing because we invoked these tests with

   python setup.py test

but the return value from this invocation method was zero
regardless of the test results.  The build process interprets
all steps that exit with zero as a successful step.

So this commit is going to break the build until we fix the
error that has existed for a number of builds.