Don't leak directories in /tmp in decrypt_and_uncompress().
This function always created a subdirectory in /tmp, and never deleted
it, but that directory is only needed for the GPG decryption path.  Instead
use os.tempnam() to create the directory name, and a try/finally block
to clean it up.  Do the same thing in compress_and_encrypt() (which didn't
have this directory leak bug).