Don't logger.warn() on ImportError; do relative import.
This pattern is generally wrong:

try:
  import foo
except ImportError:
  logger.warn('cannot import foo', exc_info=True)
  foo = None

It violates the "no side effects on import" rule, and in most real-world
application contexts, logging isn't actually set up yet, so all the end
user sees is 'No handlers could be found for logger "streamcorpus"', on
every tool invocation.

In particular, don't attempt to log a warning for backports.lzma or snappy
imports; but do throw a RuntimeError on snappy_* if it's not there.

Also in some test contexts, the sequence where streamcorpus.__init__ imports
* from streamcorpus.package_globals, which in turn tries to import
streamcorpus.things, causes a dependency-resolution loop, so switch
package_globals (and _chunk) to use relative imports (which work as of at
least Python 2.6).